Renamed Read handler to be List handler, renamed View handler to be Read handler. Updated routing and config to reflect the changes.
