+++ src/Announcements/src/ConfigProvider.php
@@ -51,11 +51,11 @@ class ConfigProvider
             'invokables' => [
             ],
             'factories'  => [
-                Handler\AnnouncementsReadHandler::class => Handler\AnnouncementsReadHandlerFactory::class,
-                Handler\AnnouncementsViewHandler::class => Handler\AnnouncementsViewHandlerFactory::class,
                 Handler\AnnouncementsCreateHandler::class => Handler\AnnouncementsCreateHandlerFactory::class,
+                Handler\AnnouncementsReadHandler::class => Handler\AnnouncementsReadHandlerFactory::class,
                 Handler\AnnouncementsUpdateHandler::class => Handler\AnnouncementsUpdateHandlerFactory::class,
                 Handler\AnnouncementsDeleteHandler::class => Handler\AnnouncementsDeleteHandlerFactory::class,
+                Handler\AnnouncementsListHandler::class => Handler\AnnouncementsListHandlerFactory::class,
             ],
         ];
     }
@@ -97,14 +97,14 @@ class ConfigProvider
             [
                 '__class__' => RouteBasedResourceMetadata::class,
                 'resource_class' => Announcement::class,
-                'route' => 'announcements.view',
+                'route' => 'announcements.read',
                 'extractor' => ReflectionHydrator::class,
             ],
             [
                 '__class__' => RouteBasedCollectionMetadata::class,
                 'collection_class' => AnnouncementCollection::class,
                 'collection_relation' => 'announcement',
-                'route' => 'announcements.read',
+                'route' => 'announcements.list',
             ],
         ];
     }
